update working with array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,6 @@
 #include "function_pointers.h"
 #include <stdio.h>
+#include <stddef.h>
 
 /**
  * array_iterator -> executes function on array argument
@@ -11,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int index = 0;
+	size_t index = 0;
 
 	if (action == NULL)
 		return;
